pyproject.toml
[tool.poetry]
name = "discord-amethyst"
version = "2.0.0"
description = "An opinionated extension for discord.py."
authors = ["NimajnebEC <nimajnebec@users.noreply.github.com>"]
packages = [{ include = "amethyst", from = "src" }]
repository = "https://github.com/NimajnebEC/discord-amethyst"
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
lavender-logging = "^1.0.0"
environ-import = "^0.1.0"
discord-py = "^2.3.2"
dynamicpy = "^1.2.2"
croniter = "^1.4.1"
python = "~3.11"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
ruff = "^0.1.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 92
